Add Frame component
This component is what is used for controlling app location/context
switching. It handles, via the fade or slide=left/right attributes, the
sliding in/out and fading in/out of a Frame based upon whether it is
marked as active or not.

This is basically the standard mobile slider functionality.